Memory optimization on cell objects.
Changed the cell source type from integer to short integer in the CLR
wrapper.
Change the source type enumeration from integer to short integer in
SharpBoostVoronoi library.

+++ SharpBoostVoronoi/BoostVoronoiTests/TestCSharp.cs
@@ -161,7 +161,7 @@ namespace BoostVoronoiTests
                 vw.AddSegment(s.Start.X, s.Start.Y, s.End.X, s.End.Y);
 
             vw.ConstructVoronoi();
-            List<Tuple<int, int, bool, bool, List<int>, bool, int>> clrCells = vw.GetCells();
+            List<Tuple<int, int, bool, bool, List<int>, bool, short>> clrCells = vw.GetCells();
 
             //Build the C# Voronoi
             BoostVoronoi bv = new BoostVoronoi();
@@ -191,7 +191,7 @@ namespace BoostVoronoiTests
                 vw.AddSegment(s.Start.X, s.Start.Y, s.End.X, s.End.Y);
 
             vw.ConstructVoronoi();
-            List<Tuple<int, int, bool, bool, List<int>, bool, int>> clrCells = vw.GetCells();
+            List<Tuple<int, int, bool, bool, List<int>, bool, short>> clrCells = vw.GetCells();
 
             //Build the C# Voronoi
             BoostVoronoi bv = new BoostVoronoi();
@@ -235,7 +235,7 @@ namespace BoostVoronoiTests
                 vw.AddSegment(s.Start.X, s.Start.Y, s.End.X, s.End.Y);
 
             vw.ConstructVoronoi();
-            List<Tuple<int, int, bool, bool, List<int>, bool, int>> clrCells = vw.GetCells();
+            List<Tuple<int, int, bool, bool, List<int>, bool, short>> clrCells = vw.GetCells();
 
             //Build the C# Voronoi
             BoostVoronoi bv = new BoostVoronoi();
@@ -275,7 +275,7 @@ namespace BoostVoronoiTests
                 vw.AddSegment(s.Start.X, s.Start.Y, s.End.X, s.End.Y);
 
             vw.ConstructVoronoi();
-            List<Tuple<int, int, bool, bool, List<int>, bool, int>> clrCells = vw.GetCells();
+            List<Tuple<int, int, bool, bool, List<int>, bool, short>> clrCells = vw.GetCells();
 
             //Build the C# Voronoi
             BoostVoronoi bv = new BoostVoronoi();
@@ -340,7 +340,7 @@ namespace BoostVoronoiTests
                 vw.AddSegment(s.Start.X, s.Start.Y, s.End.X, s.End.Y);
 
             vw.ConstructVoronoi();
-            List<Tuple<int, int, bool, bool, List<int>, bool, int>> clrCells = vw.GetCells();
+            List<Tuple<int, int, bool, bool, List<int>, bool, short>> clrCells = vw.GetCells();
 
             //Build the C# Voronoi
             BoostVoronoi bv = new BoostVoronoi();
